Move Ssl class to jdbc package for use other plugins.
